success_rate implemented

+++ dicesGame/database/migrations/2023_04_23_165436_add_success_rate_to_users_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->float('success_rate')->nullable();
+            $table->decimal('success_rate', 3, 2)->nullable();
         });
     }
